# src/openapi.yml
openapi: 3.1.0
info:
  title: Thesis Rewriting Platform API
  version: 0.1.0-beta
  description: |
    Authenticated REST contract for the current beta. Product endpoints are
    served below `/api`; Auth.js owns the `/auth` routes. Mutations may include
    `X-Mutation-Id` so clients can correlate the REST response with a committed
    Socket.IO event.
servers:
  - url: http://localhost:3001
    description: Local Express server
...
paths:
...
  /api/me:
    get:
      tags: [Authentication]
      summary: Return the authenticated Auth.js user
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Auth.js user
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: '#/components/schemas/AuthUser'
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/users/me:
    get:
      tags: [Profile]
      summary: Read the product profile and aggregate writing statistics
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Product profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductProfile'
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/users/me/activity:
    post:
      tags: [Profile]
      summary: Record one idempotent local-calendar-day activity
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [timeZone]
              properties:
                timeZone:
                  type: string
                  description: Valid IANA time zone, such as America/Toronto.
      responses:
        '200':
          description: Updated profile and streak
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductProfile' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/users/me/profile-picture:
    get:
      tags: [Profile]
      summary: Download the authenticated user's profile picture
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Image bytes
          content:
            image/png: {}
            image/jpeg: {}
            image/webp: {}
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Profile]
      summary: Replace the authenticated user's profile picture
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/MutationId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductProfile'
        '400': { $ref: '#/components/responses/BadRequest' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: authjs.session-token
      description: Auth.js database-session cookie. The exact cookie name can vary by deployment security settings.
  parameters:
    MutationId:
      name: X-Mutation-Id
      in: header
      required: false
      schema: { $ref: '#/components/schemas/Uuid' }
      description: Optional client-generated UUID echoed in the related realtime event.
...
  responses:
    BadRequest:
      description: Invalid request body, identifier, or file
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing, expired, or untrusted authenticated session
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource does not exist for the authenticated user
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    Uuid:
      type: string
      format: uuid
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code:
          type: string
          enum: [AI_NOT_CONFIGURED, AI_RATE_LIMITED, AI_TIMEOUT, AI_INVALID_OUTPUT, AI_PERSISTENCE_FAILED, STALE_BLOCK_CONTEXT, ANALYSIS_REQUIRED]
        correlationId: { type: string, format: uuid }
    AuthUser:
      type: object
      required: [id, email]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        name: { type: [string, 'null'] }
        email: { type: string, format: email }
        image: { type: [string, 'null'], format: uri }
    UserStats:
      type: object
      required: [completed_chars, total_chars, completed_rate, streak_day_count]
      properties:
        completed_chars: { type: integer, minimum: 0 }
        total_chars: { type: integer, minimum: 0 }
        completed_rate: { type: number, minimum: 0, maximum: 1 }
        streak_day_count: { type: integer, minimum: 0 }
        streak_start_date: { type: [string, 'null'], format: date }
        streak_end_date: { type: [string, 'null'], format: date }
        streak_days:
          type: array
          items: { type: string, format: date }
        last_active_on: { type: [string, 'null'], format: date }
        time_zone: { type: [string, 'null'] }
    ProductProfile:
      type: object
      required: [id, auth_user_id, email, hasProfilePicture, stats]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        auth_user_id: { $ref: '#/components/schemas/Uuid' }
        email: { type: string, format: email }
        display_name: { type: [string, 'null'] }
        profile_picture_mime: { type: [string, 'null'] }
        hasProfilePicture: { type: boolean }
        stats: { $ref: '#/components/schemas/UserStats' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...